+++ cmsafvis/R/monitor_climate.R
@@ -4,6 +4,8 @@
 #' the given variable within the given time range and area. The intended 
 #' application is for daily accumulated data, such as sunshine duration.
 #' Dependent on the output format a PNG or MP4 is created.
+#' 
+#' Circular stripe plots are inspired by Emanuele Bevacqua (see emanuele.bevacqua.eu)
 #'
 #' You can pass a YAML config file and/or specify the arguments directly.
 #' Argument prioritization is done in the following way:
@@ -54,6 +56,7 @@
 #' @param analyze_method Two analyze methods: mean == FALSE; accumulate == TRUE (logical) (Warming Stripes Plots, Time Series Plots, Trend Plots)
 #' @param selected_number Timesteps of the same selected time range (Warming Stripes Plots, Time Series Plots, Trend Plots)
 #' @param stripe_color Color option for stripe plots (Warming Stripes Plots)
+#' @param circ_plot Circular stripe plots (logical)
 #' @param dwd_logo Whether to add the DWD logo (logical).
 #' @param verbose Whether to display progress messages (logical).
 #' @param nc Alternatively to \code{infile} you can specify the input as an
@@ -96,6 +99,7 @@ monitor_climate <- function(plot_type = "absolute_map",
                             analyze_method = TRUE,
                             selected_number = 1,
                             stripe_color = 1,
+                            circ_plot = FALSE,
                             dwd_logo = FALSE,
                             verbose = TRUE,
                             nc = NULL